Remove kicad-4.0.0-nostrip.patch
It was added in 2014 without any testing to resolve a problem with an
empty kicad-debuginfo package. The first part of the patch was not
relevant for Linux anyways (MinGW only), and the debuginfo package is
just fine nowadays without the patch.
See https://bugzilla.redhat.com/show_bug.cgi?id=1076929 for details.