Move 3D models into separate RPM kicad-packages3d
The 3D models from the kicad-packages3D repository occupy multiple GB of
disk space. Users should not be forced to install them. This also allows
to avoid re-downloading the (mostly unchanged) 3D files every single day
as part of the nightly build (which is especially useful in case of a
slow Internet connection).